refactor: streamline modal import styles and improve closing animation
