+++ src/app/layout/sections/modal/modal-import/modal-import.component.css
@@ -1,60 +1,1 @@
-/* Modal Import Styles */
-
-/* Backdrop fade animation */
-@keyframes backdropFadeIn {
-  from {
-	opacity: 0;
-  }
-  to {
-	opacity: 1;
-  }
-}
-
-@keyframes backdropFadeOut {
-  from {
-	opacity: 1;
-  }
-  to {
-	opacity: 0;
-  }
-}
-
-/* Modal slide-in animation */
-@keyframes modalSlideIn {
-  from {
-	opacity: 0;
-	transform: scale(0.95) translateY(-20px);
-  }
-  to {
-	opacity: 1;
-	transform: scale(1) translateY(0);
-  }
-}
-
-@keyframes modalSlideOut {
-  from {
-	opacity: 1;
-	transform: scale(1) translateY(0);
-  }
-  to {
-	opacity: 0;
-	transform: scale(0.95) translateY(-20px);
-  }
-}
-
-/* Apply animations via CSS classes */
-.modal-backdrop-enter {
-  animation: backdropFadeIn 200ms ease-out forwards;
-}
-
-.modal-backdrop-leave {
-  animation: backdropFadeOut 150ms ease-in forwards;
-}
-
-.modal-content-enter {
-  animation: modalSlideIn 250ms ease-out forwards;
-}
-
-.modal-content-leave {
-  animation: modalSlideOut 200ms ease-in forwards;
-}
+/* Modal Import Styles - Uses global animations from assets/css/animations.css */
