--- src/app/types/index.ts
@@ -1,6 +1,0 @@
-export * from './API';
-export * from './Athlete';
-export * from './Country';
-export * from './Disciplines';
-export * from './Importer';
-
